extract boxed factory
